requirements.txt
# Basic dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and related libraries
transformers>=4.35.0
tokenizers>=0.14.0
accelerate>=0.24.0
peft>=0.7.0
bitsandbytes>=0.41.0

# Dataset processing
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0

# Reinforcement learning
trl>=0.7.0

# Evaluation metrics
scikit-learn>=1.3.0
nltk>=3.8.1
rouge-score>=0.1.2
bert-score>=0.3.13

# Configuration and logging
pyyaml>=6.0
wandb>=0.15.0
tensorboard>=2.14.0

# Utility libraries
tqdm>=4.65.0
click>=8.1.0
rich>=13.0.0

# Audio processing
librosa>=0.10.1
soundfile>=0.12.1
speechrecognition>=3.10.0

# Text processing
regex>=2023.8.8
sentencepiece>=0.1.99

# Other utilities
requests>=2.31.0
pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 开发工具 (可选)
jupyter>=1.0.0
ipywidgets>=8.0.0
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0

# GPU 支持 (可选，根据环境选择)
# torch[gpu] 或者手动安装 CUDA 版本的 PyTorch

# 额外的语言处理工具
spacy>=3.7.0
polyglot>=16.7.4
googletrans>=4.0.0

# API 和网络请求
httpx>=0.25.0
aiohttp>=3.8.0

# 数据存储
sqlite3  # Python 内置
redis>=5.0.0  # 可选，用于缓存

# 文档生成
sphinx>=7.0.0  # 可选